weavloader/requirements.txt
weaviate_client==4.14.* #https://weaviate.io/developers/weaviate/release-notes#weaviate-core-and-client-releases
imageio==2.33.1
sage_data_client==0.7.1
matplotlib==3.8.2
Pillow==10.4.*
tritonclient[grpc]==2.53.*
numpy==1.24.*
celery==5.3.*
redis==5.0.*
prometheus-client==0.19.*
flask==3.0.*
psutil==5.9.*
flower==2.0.*
Requests